Hook creation for FetchNotifications
