+++ frontend/src/hooks/useMaintenanceNotifications.ts
